backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.4
pydantic-settings==2.7.1
httpx==0.28.1
google-api-python-client==2.159.0
google-auth==2.37.0
python-dotenv==1.0.1
Pillow==11.1.0
